Fix deploy script to build Docker image locally if not in registry
The Docker images haven't been built yet since we just added the Dockerfiles.
This update allows the deployment script to automatically build the image
locally if it can't be pulled from the registry.

This fixes the deployment issue where the image pull fails because
the CI/CD hasn't run yet to build and push the images.